security: Remove TOTP and QR code
It made it too complicated to share and get access to the resources. Now, the users
just have to share an URL and a password, which is easy to understand. If the
password leaks, then an attacker can now get access to the site (before, they needed
to get access to the TOTP secret as well).